Use the global distances matrix
In order to do multidimensional arrays where it's defined in the global
space without knowing the size we need to allocate a contiguous amount
of memory for both the pointers to the rows, as well as all of the
row-column indices.

We can do this by allocating the memory for this, then going through all
of the rows and assigning the pointer location to the index.